Move wrappers to test_process_pr
